+++ test/e2e/e2e_test.go
@@ -55,7 +55,7 @@ var (
 	errTokenMissing = errors.New("token must be specified in DIGITALOCEAN_ACCESS_TOKEN environment variable")
 
 	// De-facto global variables that require initialization at runtime.
-	supportedKubernetesVersions     = []string{"1.22", "1.21", "1.20", "1.19"}
+	supportedKubernetesVersions     = []string{"1.26", "1.25", "1.24", "1.23"}
 	sourceFileDir                   string
 	testdriverDirectoryAbsolutePath string
 	deployScriptPath                string
@@ -391,9 +391,11 @@ func createCluster(ctx context.Context, client *godo.Client, nameSuffix, kubeMaj
 		Tags:        []string{"csi-e2e-test", versionTag, fmt.Sprintf("branch-identifier:%s", nameSuffix)},
 		NodePools: []*godo.KubernetesNodePoolCreateRequest{
 			{
-				Name:  clusterName + "-pool",
-				Size:  "s-4vcpu-8gb",
-				Count: 3,
+				Name:      clusterName + "-pool",
+				Size:      "s-4vcpu-8gb",
+				MinNodes:  5,
+				MaxNodes:  20,
+				AutoScale: true,
 			},
 		},
 	})